requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
pyyaml==6.0.1
numpy==1.24.3
google-genai==0.2.2

